A new opam-standard metadata file describing the alternate views on this repository
My primary concern is the requirement for behavior of these URLs. I believe that the 'browse' URL should point to the browseable package index and the 'upstream' URL should point to the primary branch's root. Perhaps 'browse' should point to a location that corresponds to the repo as a whole (incl compiler switches) and the 'upstream' should point to a branch-agnostic interface. These locations are more flexible but more onerous to use, however, as they require further metadata at those locations to get you to where you want to go.