add opam package for pfff, just the parsing libraries
Summary:
This is the Initial version of pfff opam package. It just installs
the different parsing libraries. In the next week I'll
try to update it to also install sgrep, spatch, stags,
scheck, and later codemap and codegraph. But for now,
let's just install the libraries. Js_of_ocaml recently
copy pasted the js parser in pfff just because it
was not easily accessible as a standalone library.

Test Plan:
put this package in my ~/testing repo
opam update testing
opam install pfff
 => it works, it installs pfff-xxx libraries
opam uninstall pfff
 => it works, it uninstalls pfff-xxx libraries