Experimental support for Travis source install
If a package defines a depext with the `["source" "linux"]` tags,
it will be treated as a URL and piped through a shell.  This script
should download and install any external source dependencies
required.

This depext is run *after* any binary depexts, so they can install
prerequisite libraries.

This commit includes a working sodium.0.1.0 as an example.