Heptagon/BZR version 1.00.06
Heptagon/BZR is a synchronous dataflow language whose syntax and
semantics is inspired from Lustre, with a syntax allowing the
expression of control structures (e.g., switch or mode automata).

Heptagon/BZR is a research compiler, whose aim is to facilitate
experimentation. The current version of the compiler includes the
following features:
- Inclusion of discrete controller synthesis within the compilation
- Expression and compilation of array values with modular memory optimization

See http://bzr.inria.fr for further informations.