Addition of portia, a literate programming tool.
Portia is a programmable text processor that allow to extract fragments from a
set of files and recompose some output files from them. It is meant to serve as
a literate programming preprocessor, agnostic to both the format used for the
documentation and the programming language, and thus is adapted to OCaml.

It is itself implemented using literate programming and needs funnelweb
(http://www.ross.net/funnelweb/) for bootstrap.