Allow 4.02.1+32bit to build cmxs on Darwin
This is an admittedly opaque change that allows to build cmxs' on Darwin.
The crux of the issue is that ld normally updates only vanilla relocations,
and OCaml emits instruction-based ones. On i386, this results in a linker
error; the added flag enables the linker to update all relocations
instead.