Explicitely specify binaries path in biniou
If no PREFIX or BINDIR is specified, the Makefile tries to compute a
decent one, using in particular the 'which' primitive. However, if
/bin/sh is minimal, 'which' may not exist, thus leading to PREFIX
being empty and BINDIR being /bin, which is incorrect.

As opam can provide an explicit path for BINDIR, providing it is a way
to solve the problem.