Newest packages for Apron (with OSX support) & mlgmpidl (1.2.1)
The new package `mlgmpidl' installs a binding to GNU MP and GNU MPFR
that does not match the one provided by `mlgmp'. It can also be
considered more lightweight as it does not depend on `oasis'.

The `mlgmp' package is also marked as conflicting with earlier
versions of `apron', as the `mlgmpidl' library was previously included
and installed by apron.