ocamlbuild bulk deps: add some missing packages
Guillaume Bury reported a typo in the ocamlbuild-detection logic: it
looked for a file name "_tag" instead of "_tags", and thus missed some
ocamlbuild users. The present patch is the result of re-running the
script after fixing this: 171 additional packages are detected to use
ocamlbuild (including a few maintained by Guillaume Bury).

The last commit on master at the time the script was re-run was

> commit a939e2dd452c9cd24a55a359cb09faa145f63b32
> Merge: 58e08f2 7016e0b
> Author: yallop <yallop@gmail.com>
> Date:   Tue Dec 1 23:08:32 2015 +0000
>
>     Merge pull request #5221 from gabelevi/flowtype
>
>     +flowtype.0.19.0