fix version fields of newer beta compilers
The 'version:' field in .comp files should be equal to what 'ocaml -vnum' returns, it's not the identifier
of the compiler. This causes errors that opam doesn't handle very well (this confirms that handling
compilers and packages in a uniform way was a better idea). See https://github.com/ocaml/opam/issues/2516 for an example.