Detect various configure parameters at installation time
The variables CFLAGS, LDFLAGS, CPPFLAGS and OCAMLFLAGS are deduced
from the value of the PATH when the conf-automake package installed.
These values are stored in the configuration file for the package, so
that they can be used in opam files for other packages. The guessed
value of CFLAGS is stored as

  cflags: "CFLAGS=-I..."

and so on.  It makes it especially easy to use the variable on
command lines, like

  ["env" conf-autoconf:cflags "sh" ...]

or

  ["./configure" "--prefix" prefix conf-autoconf:cflags]

for instance.