Revert "Merge pull request #6657 from gasche/ocamlfind-no-ncurses"
This reverts commit f45249c61b7f5eece77ccda6fd37acf4bab39af5, reversing
changes made to e7a191dd0d0c493002e2d80ed6a3cb6a029e295f.

I proposed to remove the ncurses dependency after:
- noticing that it caused failure for some users (eg. OpenSUSE)
  http://lambda.jstolarek.com/2016/05/installing-ocaml-under-opensuse-11-4-or-the-compilation-of-conf-ncurses-failed/
- checking with Gerd Stolpmann that ocamlfind did not depend on ncurses

It turns out that the dependency comes from a hard dependency of the
OCaml compiler runtime for ncurses, that infects any software compiled
with -custom or embedding the runtime (eg. custom toplevels). It is
wrong to have the dependency at the ocamlfind level, but removing it
revealed a similar issue for plenty of downstream packages and it
breaks lots of stuff.

Let's revert that change for now and think about the issue more
carefully. I think a better long-term solution would be to do away
with the ncurses dependency in the compiler distribution.

For more discussion, see:
- https://github.com/ocaml/opam-repository/pull/6657
- https://github.com/ocaml/opam-repository/pull/6773