archimedes: mark (< 4.03) for versions 0.{15,17}, fixed in 0.18
opam-builder report:
  <http://opam.ocamlpro.com/builder/html/archimedes/archimedes.0.4.17/33387e164895bd5768930c4d18ba1b65>

The incompatibility was caused by code of the form "let x = 2.in",
rejected by the new parsing of numeric literals that considered "2.in"
as a literal token with custom suffix.

There are other opam-builder failures for the Archimedes package on
older OCaml versions, that seem related to dependencies issue with the
Graphics module (which is only marked as an optional dependency but
may be required after all). See for example:
  <http://opam.ocamlpro.com/builder/html/archimedes/archimedes.0.4.18/3b5c9d422a00ac5da03cb25cb34718bb>

(cc: @Chris00)