Add opam-lib 1.3.0
Hopefully this can be a first, easier step to migrate the tooling to 2.0; this version is still fully compatible with the 1.2.2 format repository, but already has a few of the improvements of opam 2, and e.g. a more complete opam lint.

This is intended to be the new base for the tooling (e.g. Camelus, opam2web...) while the repository is still on opam 1.2.2; it will also, later, be needed for the 2.0.0 to 1.2.2 format downgrade tool)