Align key_schedule on a 16 byte boundary
Memory access to/from XMM registers require the memory operand to be
aligned in order to have good performance (using the `MOVDQA`
operation). To ensure that, the compiler is supposed to align stack
variables, but it can not always do that. In particular, old GCC
versions (4.4.7, present in Centos 6) fail to do that. It is thus
necessary to align it manually using extra space.