package the last release of ppx_tools, with 4.05 support
Absence of ppx_tools has been a major source of breakage in the 4.05
opam repository so far. This should unlock the test of many more opam
packages during the beta period.

(Thanks to Alain Frisch for his extremely quick release of a new
ppx_tools versions when I raised the issue.)