capnp: do not use latest Core until OCaml 4.04.1 comes out
There is a bug in OCaml 4.04.0 that causes anything that links with
Ephemeron to fail in the toplevel, since it was accidentally missed
in the release from the stdlib/StdlibModules file.

This causes capnp builds to fail with sexplib>=v0.9.0 which does
link to it. The constraint here is not quite right, since it should
all work when ocaml 4.04.1 is released, but we cannot express the
constraint easily.

So for now, removing the build failure and will figure out the
compiler constraint when the new version of ocaml is released.

thanks @diml for diagnosing the issue