ADD Add opass 2.15
This is a backwards breaking update of opass.  The CLI has been changed from
Core's parser to Cmdliner.  Cmdliner does not support single-dash parameters
that are more than a character long.  With that, all of the parameters such as
"-copy" and "-show" are now "--copy" and "--show".  However, many parameters
have a single character option as well, such as "-c" and "-s".

The database format is unchanged.

A new search parameter, "-p" or "--pattern", has been added where the search
terms can be a Lua pattern.