Remove old xapi packages
No package is depending on them and they are unmaintained and probably not compiling.
We are re-constructing a working opam repository in https://github.com/xapi-project/xs-opam
and we will start updating and maintainig the opam files of the stable libraries.
However some of them, like xapi, xapi-networkd, xapi-idl, ... are always in motion and
we have very strict interdepenencies: this makes it very hard to guarantee that the packages
constraints are sound to be in the opam repository and their presence in the opam repository
makes it impossible to manage a switch containing both opam and xs-opam without pinning.

Removing these libraries is the first step towards being able to maintain and update the ones
that are remining in opam.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>