+tcpip-3.1.3
Bugfix release with more correct handling of the interaction between
IP-module expressed maximum transmission unit (MTU) and the TCP maximum
segment size (MSS).  This release is preferred over 3.1.0 and 3.1.1
(which would crash on some backends under some circumstances) and 3.1.2
(which artificially pinned the maximum segment size at a common, but not
universal, value).

tcpip versions prior to 3.1.0 did not support TCP decision-making based
on the MTU and had behavior similar to 3.1.2.