Add "num" package for the Num library
The Num library for arbitrary-precision integer and rational arithmetic has been part of the OCaml distribution since the first public release of OCaml.  It is scheduled to be removed from the OCaml distribution in version 4.06 of OCaml, at which time the Num library will remain available as a separate package.

In preparation for this transition, this commits adds a "num" package and a "base-num" package.
- "base-num" is a dummy package available if OCaml version < 4.06
- "num" has an empty version 0 implementation that just depends on "base-num"

A future commit will add a "num 1.0" implementation for OCaml 4.06 and up.