tcpip: upper bound on mirage-profile
this is due to the removal of an accidental cstruct.ppx linkage
in mirage-profile, which means that downstream packages that
depended on that now fail to build.  there is an associated release
of tcpip coming that fixes this issue in that package as well.

see #9543 revdeps and https://github.com/mirage/mirage-tcpip/pull/318