nocrypto: make lwt a firm opam dependency of nocrypto
this removes the depopt from opam, but lwt remains optional from
a findlib perspective. The reason this is necessary is to allow
cstruct-lwt to work (>=cstruct.3.0) which is difficult to do if
lwt remains a depopt.

An upstream release will be necessary to fully support cstruct3+
while lwt remains a depopt, but this unblocks the conflict in the
stable opam repository for now when using any version of cstruct