postgresql: add some depexts
Besides postgresql client libraries, this package also relies on the
'pg_config' executable. Some package systems distribute this as part
of their "postgresql-dev{,el}" package, but it may also be in
"postgresql" itself, and it is not easy to see whether it is present
on a given platform -- it requires a searchable list of files. I did
not try to fill BSD and OSX depexts for this reason.