ppx_deriving 4.2.1
This is a minor release on top of 4.2 (instead of a minor release on
top of 4.1 as was originally discussed). The goal is to release this
first, ask user feedback on whether it solves their issues, and make
a 4.1 maintenance release if the issues with driverification are still
there.

The changelog is as follows

  * Add support for OCaml 4.06.0
    ocaml-ppx/ppx_deriving#154, ocaml-ppx/ppx_deriving#155, ocaml-ppx/ppx_deriving#156, ocaml-ppx/ppx_deriving#159
    (Gabriel Scherer, Fabian, Leonid Rozenberg)
  * Consider { with_path = false } when printing record fields
    ocaml-ppx/ppx_deriving#157
    (François Pottier)