+yaml.0.1.0
This is an OCaml library to parse and generate the YAML file format.  It is
intended to interoperable with the [Ezjsonm](https://github.com/mirage/ezjsonm)
JSON handling library, if the simple common subset of Yaml is used.  Anchors
and other advanced Yaml features are not implemented in the JSON compatibility
layer.