Big new year cleanup
This is probably not to be merged as is, but may be useful to open some discussions, and make some benches and tests.

What this patch does is remove any package that
- has a version in the form 'MAJOR.MINOR.PATCH'
- is the non-latest patch revision of the non-latest minor revision

In other words, we keep only the latest patch revision of all versions but the current one.

For example, if you have 1.1.0, 1.2.1, 1.2.2, 1.3.0 and 1.3.1, only 1.2.1 will be removed.