Warn if num installed to a broken system compiler
If the num library was installed in a system switch where the user had
write permissions to the system compiler's lib directory before #11207
was merged, then *fresh* installations of the ocamlfind package will
detect this "system" installation of num and create a META file for it.
This in turn will cause the num package installation to fail.
Annoyingly, this failure only appears once because opam then executes
ocamlfind remove num so a subsequent attempt to install the num package
will appear to have worked.

This patch displays a comprehensive error message the first time,
strongly suggesting that the user manually delete the files which were
incorrectly installed previously.