restores constraints for bap-elf and drops bap-elf.1.4
Probably we didn't merge back the constraints for bap-elf, or somehow
messed with the merge. Fortunately, the CI captured this.

Concerning the bap-elf.1.4, we are skipping it, as it doesn't compile
with this version due to:

- https://github.com/BinaryAnalysisPlatform/bap/issues/756
- https://github.com/janestreet/ppx_driver/issues/14

We need to wait until the latest ppx_driver is released and also
update our suite to core v0.10, that won't happen until the 1.5
release.