makes find-ida to soft-fail if IDA is not found
So far, we didn't allow the opam users to install bap-ida if conf-ida
is not installed. That's usual for opam packages that somehow rely on
external system packages on availability of other software on a
system.

Our IDA integration is a little bit different, as it allows
specification of IDA path later, moreover it supports multiple IDA
installations. So, the conf-ida should not imply such strict policies
and prevent people from installing IDA integration even when IDA is
not present (or not foundable - e.g. when it is containerized in
Docker).