afl: set AFL_NO_X86=1 for build/install
AFL_NO_X86 allows afl's build to skip a test for X86 that's relevant for
testing whether `afl-gcc` works on that architecture, but not for the
workflow most likely for OCaml users (where instrumentation is generated
by `ocamlopt` and more closely resembles the "llvm mode" in AFL's
documentation).  This test fails on non-x86 architectures, so set
AFL_NO_X86 to disable it.

For more information, see `docs/INSTALL` in afl's source tree.