flowtype.0.27.0: restrict to os != linux
In bulk builds, this particular (historical) version of Flow
causes a timeout in builds as it seems to be in an infinite loop
in a Makefile.  The problem is fixed in later versions, so it is
just this one version affected, and is only happening on Linux
toolchains.

This PR just disables Linux support, which will allow the old
version to be built by disabling that constraint, but also unblocks
CI resources in the short term.
t