conf-pkg-config: Install pkg-config symlink on OpenBSD
The /usr/bin/pkg-config provided by OpenBSD does not support
${pcfiledir}, which breaks .pc files used by MirageOS. The OpenBSD
devel/pkgconf package does support this feature, but does not install
itself as "pkg-config" on the PATH. This change adds a symlink to the
correct "pkg-config" in the in the OPAM "bin" directory at package
installation time.

We also inform the user, since this will change the "pkg-config" binary
on their PATH when using OPAM..

Additionally, we mark conf-pkg-config.1.0 as unavailble on OpenBSD to
prevent installation of the buggy version.