mirrors the FrontC package
The upstream package is no longer available, so as a temporary measure I propose to mirror it (thanks @pmetzger for the archive). We have contacted the upstream author, and if he won't reply or if he will refuse to fix, then we will put the package on a more reliable media (e.g., github).