merge descr and url files into opam files
Some descr and url files have been merged in the
days after the merge to opam2. This PR is to
make it right and include them in the opam file
instead.

A forthcoming PR will add a synopsis or description
to the packages that are currently missing one
when it is possible to get it.

Signed-off-by: Marcello Seri <marcello.seri@gmail.com>