New package: conf-libclang
This configuration package will enable the release of the next version
of clangml.

This package checks whether there exists one working version of
libclang and llvm (by opposition to conf-llvm, that checks whether one
specific version is available). This behaviour is linked to the fact
that the next version of clangml will be able to compile with any
version of libclang.