Restrict ppx_deriving_yojson dependency to yojson < 1.6
Yojson 1.6 deprecates the type `json` in favor of the new type `t`.
This shouldn't be an issue but users of ppx_deriving_yojson and
dune are gonna have a hard time working with yojson 1.6 since warning 3
is fatal in the default dune profile and they can't change
ppx_deriving_yojson output.