ppx_cstruct: tests fail under 4.08.0
With 4.08 the compiler prints errors differently by default,
which appears to break expect-style tests that are checking for errors.

This should probably be fixed at the upstream level of dune (probably
by calling the compiler with the '-error-style short' option in
tests), so in the future (if "fixed" versions of dune are released)
the (ocaml:version < 4.08) constraint on tests should be replaced by
(dune:version >= fixed-version) to re-enable testing.