Fix opam reinstall graphics
If the package is reinstalled, then it will be compiled while
graphics.cmi is already in-place, which means that the existing test
would assume the package was already built.

The package now installs a status file in
$SWITCHROOT/share/graphics/state which if found is used to determine if
the detected graphics.cmi was in the switch from a previous build or
from is from the compiler itself.

If the underlying compiler is reinstalled, then opam would remove the
graphics package before recompiling and so the state file will not exist
when graphics is built and detection will be repeated, thus this
sequence:

  sudo apt remove libx11-dev
  opam switch create test ocaml-base-compiler
  sudo apt install libx11-dev
  opam install graphics
  opam reinstall ocaml-base-compiler

will correctly reinstall the "dummy" version of this package.