conf-libev: try to build without any flags first
Before adding -I and -L flags, try to build using just $CC test.c -lev.
This will succeed if libev is installed already in the compiler's
default search path. This improves libev detection with musl-gcc, as the
musl libev is typically installed in musl-gcc's search path. With
musl-gcc, immediately searching for ev.h and adding -I and -L is likely
to find a glibc (system) libev, rather than the one compiled for use
with musl-gcc.

See

  https://github.com/ocsigen/lwt/issues/718#issuecomment-535053644