Make jbuilder.transition unavailable
This transition version was added back when dune was compatible with
jbuilder, so it made sense to install dune in all cases. However, recent
versions of dune no longer support jbuilder, so using this package just
forces installation of an old version of dune, conflicting with packages
that require a recent dune, which is generally undesirable.