remove conf-gcc from zarith
This was added as it is needed by nix, but it adds an unnecessary
dependency on gcc for other platforms where clang is the default.
We can assume there is a `cc` in the environment if opam is initialised,
so the Nix packages should be able to do it there.

cc @bobot