Add a GitHub Action to close stale issues
At the OCaml core developers meeting in April 2020, we decided
to experiment with closing stale issues that have no activity for
a long period of time. This is simply to keep the working set of
issues at mangeable levels.

This PR experiments with adding this to opam-repository, where
stale issues are particularly prone to being closable as many
have been fixed by subsequent metadata improvements. Feedback
welcome on the timeouts, and the message itself.