enables installation of the executable fixes uninstallation
It looks like the 4.8.8 version is no longer uninstalling anything, so
I've added the `remove` stanza.

I also add the installation of the `z3` executable so that we can use
`z3` without having to install the system version of z3 which could be
of different version and have other conflicts with the opam-installed one.