ocaml: compile with -fcommon to support gcc 10
gcc 10 and upwards alters the defaults such that duplicate
C symbols results in a linker error.  This PR forces compilation
with the `-fcommon` flag, which preserves the pre-gcc-10
behaviour and thus compatibility with OCaml <4.09.1.

`-fcommon` has been supported from gcc 4.4 upwards, which is
a reasonable base default (checked by @Octachron)