ocaml-variants: compile with -fcommon to support gcc10
As #16583 explains, gcc 10 and upwards alters the defaults such that
duplicate C symbols results in a linker error. This PR forces
compilation with the -fcommon flag, which preserves the pre-gcc-10
behaviour and thus compatibility with OCaml <4.09.1.

-fcommon has been supported from gcc 4.4 upwards, which is a reasonable
base default.

This patch skips a few compiler variants that already have patches,
such as termux and the bucklescript variants. Those can deal with gcc10
on their own if they still need it.

The patches are copies of the respective base patch for that compiler,
found from the ocaml-base-compiler package from #16583.