conf-binutils 0.3
This PR introduces a new version of conf-binutils and constrains all bap
packages to the old ones. The only package beyond bap universe that depends on
`conf-binutils` is `clangml`, and looks like it doesn't depend on our interface, so we leave it untoched.

What's new:

We search for files that have `cxxfilt`, `objdump`, and `readelf` suffixes in the `brew` cellar, if one available, and in the `PATH` constituents. Instead of using tools from `findutils`, `which`, and `mdfind`, as we did before, we search for the matching files manually, using the OCaml `Sys.readdir` function.

The problem is that on macOS with `brew` we can't use which utility for the search: `binutils` formula is marked as `keg-only` and also can shadow system files, so even `brew link --force` doesn't link `binutils` somewhere in the `$PATH`. That's why we have problems with installing bap on macOS: `which gobjdump` doesn't return a path to the file anymore. With this PR this problem will be solved.

Previously, we were creating two `opam` config variables per each tool, e.g. `objdump` and `objdump-targets`. But since all that we do with them on the bap side is concatenating them in one list, there is no need in several variables for one tool. Now we add only one, e.g `objdumps` or `cxxfilts` that contains a list of absolute paths to all discovered files for the given tool. The preferred tool still can be set via `opam` config or environment variable. In this case, it will be first in the paths list.